- Basic tuple support - Move filter into query grain
