+++ tests/complex/test_aggregate_of_aggregate.py
@@ -56,4 +56,4 @@ select
     generator = BigqueryDialect()
     sql = generator.compile_statement(query)
     assert "count(posts.`id`) as `default_user_post_count`," in sql
-    assert "avg(cte_posts_at_user_id_" in sql
+    assert "avg(cte_posts_at_default_user_id_" in sql
