+++ tests/stack_overflow/so_concepts/badge.preql
@@ -1,0 +1,17 @@
+import user as user;
+
+key id int;
+property id.name string;
+property id.awarded_date date;
+property id.class int;
+property id.tag_based bool;
+
+datasource badges (
+    id: id,
+    name: name,
+    date: awarded_date,
+    user_id: user.id,
+    class: class,
+    tag_based: tag_based
+    )
+address bigquery-public-data.stackoverflow.badges;
