+++ tests/stack_overflow/so_concepts/user.preql
@@ -1,0 +1,16 @@
+## users
+
+key id int metadata(description="the description");
+key location string metadata(description="The user location");
+property id.display_name string metadata(description="The display name ");
+property id.about_me string metadata(description="User provided description");
+
+datasource users (
+    id: id,
+    display_name: display_name,
+    about_me: about_me,
+    location:location
+    )
+    grain(id)
+    address bigquery-public-data.stackoverflow.users
+;
