+++ tests/stack_overflow/stackoverflow.preql
@@ -1,0 +1,8 @@
+import so_concepts.badge as badge;
+import so_concepts.question as question;
+import so_concepts.user as user;
+import so_concepts.answer as answer;
+import so_concepts.user_metrics as user_metrics;
+import so_concepts.tag as tag;
+
+
